chore(login): refactor the types package for users
