+++ frontend/src/api/user/crud/getAll.ts
@@ -2,7 +2,7 @@ import axios from 'api';
 import { ErrorResponseHandlerV2 } from 'api/ErrorResponseHandlerV2';
 import { AxiosError } from 'axios';
 import { ErrorV2Resp, SuccessResponseV2 } from 'types/api';
-import { PayloadProps, UserResponse } from 'types/api/user/listUsers';
+import { PayloadProps, UserResponse } from 'types/api/user/crud/listUsers';
 
 const listUsers = async (): Promise<SuccessResponseV2<UserResponse[]>> => {
 	try {
