+++ frontend/src/container/OrganizationSettings/InviteUserModal/InviteUserModal.tsx
@@ -1,7 +1,7 @@
 import { Button, Form, Modal } from 'antd';
 import { FormInstance } from 'antd/lib';
-import sendInvite from 'api/user/invites/create';
-import getPendingInvites from 'api/user/invites/getAll';
+import sendInvite from 'api/user/invite/create';
+import getPendingInvites from 'api/user/invite/getAll';
 import ROUTES from 'constants/routes';
 import { useNotifications } from 'hooks/useNotifications';
 import { useAppContext } from 'providers/App/App';
@@ -14,7 +14,7 @@ import {
 } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useQuery } from 'react-query';
-import { PayloadProps } from 'types/api/user/getPendingInvites';
+import { PayloadProps } from 'types/api/user/invite/getPendingInvites';
 import { ROLES } from 'types/roles';
 
 import InviteTeamMembers from '../InviteTeamMembers';
