+++ frontend/src/pages/SignUp/SignUp.tsx
@@ -2,8 +2,8 @@ import { Button, Form, Input, Typography } from 'antd';
 import logEvent from 'api/common/logEvent';
 import loginApi from 'api/login/login';
 import signUpApi from 'api/login/signup';
-import acceptInvite from 'api/user/invites/accept';
-import getInviteDetails from 'api/user/invites/get';
+import acceptInvite from 'api/user/invite/accept';
+import getInviteDetails from 'api/user/invite/get';
 import afterLogin from 'AppRoutes/utils';
 import WelcomeLeftContainer from 'components/WelcomeLeftContainer';
 import ROUTES from 'constants/routes';
@@ -15,8 +15,8 @@ import { useQuery } from 'react-query';
 import { useLocation } from 'react-router-dom';
 import { SuccessResponseV2 } from 'types/api';
 import APIError from 'types/api/error';
-import { InviteResponse } from 'types/api/user/getInviteDetails';
-import { LoginPrecheckResponse as LoginPrecheckPayloadProps } from 'types/api/user/loginPrecheck';
+import { LoginPrecheckResponse as LoginPrecheckPayloadProps } from 'types/api/login/loginPrecheck';
+import { InviteResponse } from 'types/api/user/invite/getInviteDetails';
 
 import { ButtonContainer, FormContainer, FormWrapper, Label } from './styles';
 import { isPasswordNotValidMessage, isPasswordValid } from './utils';
