+++ frontend/src/types/api/user/crud/deleteUser.ts
