+++ frontend/src/types/api/user/invite/acceptInvite.ts
