+++ frontend/src/types/api/user/invite/deleteInvite.ts
