+++ frontend/src/types/api/user/invite/getInviteDetails.ts
